Version 2
This now adds the URL of the change into the print. The next thing to do is to have a cleaner print like "3 changes to x" and "10 PushRequests to y". I don't know how to do this yet, but I guess it will be adding everything into a list, sorting and then printing.