Final version
It doesn't quite say everything that I want it to, but it does a good enough job. As of now, I am still unfamiliar with the correct verbs to add. E.g. "pushing" commits, OK, but what is the correct verb for Pull requests? This would be fairly easy to fix, so I think I am ready for the next challenge.